Updated register & auth pages

+++ src/components/Auth/Register.jsx
@@ -1,42 +1,48 @@
-import React, {useContext, useState} from 'react'
+import React, {useState, useContext} from 'react'
 import styles from './AuthRegister.module.css'
 import Footer from "../common/Footer/Footer"
 import girlImg from "../../static/girl.svg"
-import {json, useNavigate} from "react-router"
+import {useNavigate} from "react-router"
 import {UserContext} from "../../user-context"
 
-const Auth = () => {
+const Register = () => {
     const [userName, setUserName] = useState("")
+    const [email, setEmail] = useState("")
     const [password, setPassword] = useState("")
+    const [confirmPassword, setConfirmPassword] = useState("")
     const navigate = useNavigate()
-
     const {data, toggleData} = useContext(UserContext)
 
-    const signUpHandler = () => {
-        navigate("/register")
+    const signIpHandler = () => {
+        navigate("/auth")
     }
-    const signInHandler = () => {
+    const signUpHandler = () => {
         const requestOptions = {
             method: 'POST',
             mode: 'cors',
             headers: {'Accept': 'application/json', 'Content-Type': 'application/json'},
-            body: JSON.stringify({ username: userName, password: password})
+            body: JSON.stringify({ username: userName, email: email, password: password})
         }
-        fetch('http://localhost:8080/auth/signin', requestOptions)
+        if(password === confirmPassword){
+        fetch('http://localhost:8080/auth/signup', requestOptions)
             .then(response => response.json())
             .then(data => {
                 localStorage.setItem('data', JSON.stringify(data))
                 toggleData(data)
             })
+        }
         setUserName("")
+        setEmail("")
         setPassword("")
+        setConfirmPassword("")
+
         navigate("/promotions")
     }
 
     return (
-        <div className={styles.auth_body}>
-            <div className={styles.auth_wrap}>
-                <div className={styles.title_block}>
+        <div className={styles.register_body}>
+            <div className={styles.register_wrap}>
+            <div className={styles.title_block}>
                     <img className={styles.girl_img} src={girlImg} alt={"Auth Img"}/>
                     <label className={styles.logo_title}>MarlborSoft</label>
                 </div>
@@ -45,13 +51,19 @@ const Auth = () => {
                         <input className={styles.user_name_input} type={"text"} placeholder={"USERNAME"} onChange={e => setUserName(e.target.value)} value={userName}/>
                     </div>
                     <div>
+                        <input className={styles.email_input} type={"email"} placeholder={"EMAIL"} onChange={e => setEmail(e.target.value)} value={email}/>
+                    </div>
+                    <div>
                         <input className={styles.password_input} type={"password"} placeholder={"PASSWORD"} onChange={e => setPassword(e.target.value)} value={password}/>
                     </div>
                     <div>
-                        <button className={styles.auth_sign_in_btn} onClick={signInHandler}>SIGN IN</button>
+                        <input className={styles.password_input}  type={"password"} placeholder={"CONFIRM PASSWORD"} onChange={e => setConfirmPassword(e.target.value)} value={confirmPassword}/>
                     </div>
                     <div>
-                        <button className={styles.auth_sign_up_btn} onClick={signUpHandler}>SIGN UP</button>
+                        <button className={styles.register_sign_up_btn} onClick={signUpHandler}>SIGN UP</button>
+                    </div>
+                    <div>
+                        <button className={styles.register_sign_in_btn} onClick={signIpHandler}>SIGN IN</button>
                     </div>
                 </div>
             </div>
@@ -59,5 +71,4 @@ const Auth = () => {
         </div>
     )
 }
-
-export default Auth
+export default Register
